+++ test/javascript/jsClient-spec.js
@@ -7,7 +7,7 @@ describe('RAML to jsClient', function () {
   var generator = require("../../lib/generators/javascript/jsClient");
 
   var gatitosAPI = function(done){
-    test(generator, done, {package: 'org.gex.client', importPojos: 'com.pojos'}, "GatitosAPI.java", "v1/GatitosAPI.java")
+    test(generator, done, {package: 'org.gex.client', importPojos: 'com.pojos'}, "javascript/gatitosApiClient.js", "v1/gatitosApiClient.js", false)
   };
 
   it('should generate a js client from RAML file', gatitosAPI );