+++ ReleaseNotes.md
@@ -2,7 +2,7 @@
 Changing version values to RAML spec version supported, adding support to generate a js client
 
 - Moving retrofit client to src/generators/java this could break some already setup projects
-- Generating jsClient wrapping  [raml-client-generator](https://github.com/mulesoft/raml-client-generator)
+- Now it generates jsClient it uses [raml-client-generator](https://github.com/mulesoft/raml-client-generator)
 - Updating data2code to 0.0.3
 
 ## 0.0.37